+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
-All notable changes to this package will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
-and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## [2.0.4] - 2019-08-09
+### Added
+- Add related test packages
+- Add support Secondary Texture Module in Sprite Editor Window
+
+### Fixed
+- Texture and SpriteLibraryAsset subassets in PSDImporter now follows the main asset's name.
 
 ## [2.0.3] - 2019-07-20
 ### Changed